video-ingestion/requirements.txt
opencv-python
kafka-python
numpy
ultralytics
mss
requests